Add files via upload
Improved menu and menu_cursor. menu has a lot more functionality but should still work with older video series--please let me know if you run into any errors with it. menu_cursor now extends Sprite2D so that it does not interfere with Container parents, and I also added a toggle for it to follow the viewport.